requirements.txt
Flask==2.3.2
pandas==2.0.3
flask-cors==4.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
openpyxl==3.1.2